requirements.txt
pandas==2.2.0
requests==2.31.0
textblob==0.18.0
textstat==0.7.3
language_tool_python==2.7.3
scikit-learn==1.4.0
sentence-transformers==2.5.1
datasets==2.18.0
accelerate==0.29.3
matplotlib==3.8.4
seaborn==0.13.2
evaluate==0.4.1
cupy-cuda12x==13.1.0
empath==0.89
spacy==3.7.4
xgboost==2.0.3
statsmodels==0.14.2
optuna==3.6.1
pathos==0.3.2
shap==0.45.1
